backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
httpx==0.28.1
pydantic==2.12.5
python-multipart==0.0.21

# AI & RAG Dependencies
chromadb==1.4.1
sentence-transformers==5.2.0

# Groq API (FREE and FAST!)
groq==0.4.2

pypdf==6.6.0
python-docx==1.2.0
pydantic-settings==2.12.0